% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_code_perc.R
\name{step_top_code_perc}
\alias{step_top_code_perc}
\title{Top code numeric features based on percentile}
\usage{
step_top_code_perc(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  ref_val = NULL,
  prob = 0.98,
  skip = FALSE,
  id = recipes::rand_id("top_code_perc")
)
}
\arguments{
\item{prob}{A float providing the percentile to top code at}
}
\description{
\code{step_top_code_perc} creates a \emph{specification} of a recipe
step that will top code numeric data using a percentile learned
on a training set.
}
